requirements.txt
discord.py
mercadopago
qrcode
Pillow
quart==0.18.4
werkzeug==2.3.7
cryptography
pymysql